fix meeting scheduler default time bug, add backend input validation to account controller

+++ application/backend/services/AccountService.ts
@@ -1,8 +1,8 @@
 import { AccountRepository } from "./../repositories/AccountRepository";
 import { hash } from "bcryptjs";
 import { UserRepository } from "./../repositories/UserRepository";
-import { RatingRepository } from "repositories/RatingRepository";
-
+import { RatingRepository } from "./../repositories/RatingRepository";
+import { validateFields } from "./../utils/validationUtils";
 /**
  * this class encapsulates acc related business logic,
  * it interfaces between the controller and the data access layer i.e our repositories
@@ -21,6 +21,7 @@ export class AccountService {
   ) {
     this.accountRepository = accountRepository;
     this.userRepository = userRepository;
+    this.ratingRepository = ratingRepository;
   }
 
   /**
@@ -31,7 +32,7 @@ export class AccountService {
   public async getAccount(id: number) {
     const account = await this.accountRepository.getAccountById(id);
     if (!account) {
-      throw new Error("Account not found");
+      throw new Error("404");
     }
     return account;
   }
@@ -45,14 +46,12 @@ export class AccountService {
     if (!userId) {
       throw new Error("User not logged in.");
     }
-
-    try {
-      const message = await this.accountRepository.deleteAccount(userId);
-      return message;
-    } catch (error) {
-      console.error("Error deleting account:", error);
-      throw new Error("Error deleting account. : " + error.message);
+    const account = await this.accountRepository.getAccountById(userId);
+    if (!account) {
+      throw new Error("404");
     }
+    const message = await this.accountRepository.deleteAccount(userId);
+    return message;
   }
 
   /**
@@ -63,7 +62,7 @@ export class AccountService {
   public async getAccountDetails(id: number) {
     const account = await this.accountRepository.getAccountDetails(id);
     if (!account) {
-      throw new Error("Account not found");
+      throw new Error("404");
     }
     const rating = await this.ratingRepository.getUserRatingById(id);
     return { ...account, rating };
@@ -81,6 +80,13 @@ export class AccountService {
     lastName: string,
     profilePicture: string,
   ) {
+    const user = await this.accountRepository.getAccountById(id);
+    if (!user) {
+      throw new Error("404");
+    }
+    if (!validateFields({ firstName, lastName })) {
+      throw new Error("400");
+    }
     this.accountRepository.updateAccountDetails(
       id,
       firstName,
@@ -111,28 +117,37 @@ export class AccountService {
   ) {
     const existingUser = await this.userRepository.getUserById(userId);
     if (!existingUser) {
-      throw new Error("User not found.");
+      throw new Error("404");
     }
-    try {
-      let hashedPassword: string;
-      if (newPassword) {
-        hashedPassword = await hash(newPassword, 10);
-      }
-      await this.accountRepository.updateAccountAndUserDetails(
-        userId,
+
+    if (
+      !validateFields({
         firstName,
         lastName,
         email,
-        hashedPassword,
-        profilePicture,
-        weekavailability,
-        introvert,
-        isHidden,
+        newPassword,
         biography,
-      );
-    } catch (error) {
-      console.error("Error updating account:", error.message);
-      throw new Error("Error updating account. : " + error.message);
+        weekavailability,
+      })
+    ) {
+      throw new Error("400");
     }
+    let hashedPassword: string;
+    if (newPassword) {
+      hashedPassword = await hash(newPassword, 10);
+    }
+
+    await this.accountRepository.updateAccountAndUserDetails(
+      userId,
+      firstName,
+      lastName,
+      email,
+      hashedPassword,
+      profilePicture,
+      weekavailability,
+      introvert,
+      isHidden,
+      biography,
+    );
   }
 }
